add authoriztion to server and client

+++ src/features/chores/actions/delete-chore.ts
@@ -4,12 +4,28 @@ import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
 
 import { setCookieByKey } from '@/actions/cookies';
-import { fromErrorToActionState } from '@/components/form/utils/to-action-state';
+import {
+  fromErrorToActionState,
+  toActionState,
+} from '@/components/form/utils/to-action-state';
+import { getAuthOrRedirect } from '@/features/auth/queries/get-auth-or-redirect';
+import { isOwner } from '@/features/utils/is-owner';
 import prisma from '@/lib/prisma';
 import { choresPath } from '@/paths';
 
 export const deleteChore = async (id: string) => {
+  const { user } = await getAuthOrRedirect();
   try {
+    const chore = await prisma.chore.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!chore || !isOwner(user, chore)) {
+      return toActionState('ERROR', 'Not authorized');
+    }
+
     await prisma.chore.delete({
       where: { id },
     });
